Rename /shared/test-typings to /shared/test-types/
