+++ server/src/test/test-settings/testSettingsRepository.ts
@@ -1,7 +1,7 @@
 import { logger } from "server/utils/logger";
-import { TestSettings } from "shared/test-typings/models/testSettings";
+import { TestSettings } from "shared/test-types/models/testSettings";
 import { TestSettingsModel } from "server/test/test-settings/testSettingsSchema";
-import { PostTestSettingsRequest } from "shared/test-typings/api/testSettings";
+import { PostTestSettingsRequest } from "shared/test-types/api/testSettings";
 import { TestSettingsDoc } from "server/types/testSettingsTypes";
 import {
   Result,
