+++ server/src/test/test-settings/testSettingsSchema.ts
@@ -1,5 +1,5 @@
 import mongoose from "mongoose";
-import { TestSettings } from "shared/test-typings/models/testSettings";
+import { TestSettings } from "shared/test-types/models/testSettings";
 
 interface TestSettingsDoc extends TestSettings, mongoose.Document {}
