+++ server/src/types/testSettingsTypes.ts
@@ -1,4 +1,4 @@
 import mongoose from "mongoose";
-import { TestSettings } from "shared/test-typings/models/testSettings";
+import { TestSettings } from "shared/test-types/models/testSettings";
 
 export interface TestSettingsDoc extends TestSettings, mongoose.Document {}
